requirements.txt
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
idna==2.10
Pygments==2.10.0
python-dotenv==0.17.1
requests==2.25.1
requests-toolbelt==0.9.1
rich==10.12.0
typing-extensions==3.10.0.2
urllib3==1.26.5
virl2-client==2.2.1.post1
